link user and notes

+++ src/resolvers/mutation.js
@@ -1,18 +1,31 @@
+const mongoose = require('mongoose')
 const bcrypt = require('bcrypt')
 const jwt = require('jsonwebtoken')
-const { AuthenticationError } = require('apollo-server-express')
-// ForbiddenError
+const { AuthenticationError, ForbiddenError } = require('apollo-server-express')
 const gravatar = require('gravatar')
 require('dotenv').config()
 
 module.exports = {
-  newNote: async (parent, args, { models }) => {
+  newNote: async (parent, args, { models, user }) => {
+    if (!user) {
+      throw new AuthenticationError('You must be signed in to create a note')
+    }
     return await models.Note.create({
       content: args.content,
-      author: 'Adam Scott',
+      author: new mongoose.Types.ObjectId(user.id),
     })
   },
-  deleteNote: async (parent, { id }, { models }) => {
+  deleteNote: async (parent, { id }, { models, user }) => {
+    if (!user) {
+      throw new AuthenticationError('You must be signed in to delete the note')
+    }
+
+    const note = await models.Note.findById(id)
+
+    if (note && String(note.author) !== user.id) {
+      throw new ForbiddenError("You don't have permissions to delete the note")
+    }
+
     try {
       await models.Note.findOneAndRemove({ _id: id })
       return true
@@ -20,7 +33,17 @@ module.exports = {
       return false
     }
   },
-  updateNote: async (parent, { content, id }, { models }) => {
+  updateNote: async (parent, { content, id }, { models, user }) => {
+    if (!user) {
+      throw new AuthenticationError('You must be signed in to update a note')
+    }
+
+    const note = await models.Note.findById(id)
+
+    if (note && String(note.author) !== user.id) {
+      throw new ForbiddenError("You don't have permissions to update the note")
+    }
+
     return await models.Note.findOneAndUpdate(
       {
         _id: id,
